requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
groq>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
